Split AuthProvider responsibility in AuthLayout and topmost identity
AuthProvider was mixing React state and api entry point and methods,
while being outside any Router.

This complexified the management of navigation and redirection,
and the methods intended to be a actions and loaders were buggy.

Instead, the "identity" state has been brought outside of AuthProvider,
which becomes AuthLayout, an easy-going layout setting interceptors to
axios.

The previous actions and loaders defined within AuthProvider
are now standalone function that are using callbacks passed down by
RenderRoot. Therefore the whole application is still updated upon
an identity change, but the actions and loaders are now longer dependent
on it.