Add complete testsuite for Category and fixup bugs.
Fixed many bugs relative to the /api/practice endpoint,
specifically to how Category were handled and sanitized.

Testsuite passes at 100%, and a database for testing purpose
only was set up, so that the tests run in an isolated environment.

The errors message and status have been properly formalized as well.