Transition read-only ops to using lean() and simplify URI format.
This patch corrects previous blunder of always using the heavy
.exec() function upon accessing a mongoose Document.

This patch lift some of the load when obviously read-only
databases operations were spotted, so as to improve performance.

Do note that no profiling was done, and an appropriate comparison
should be carried to conclude on the performance gains.

They are however unequivocal improvements.

Signed-off-by: benjamin priour <priour.be@gmail.com>