Merge from branch fix_up_mess.
Client and servers are now both stable and tested.

Exercise creation and partial update functionalities
have been added.
Display of data under Categories have been improved.
Authentification is now persistent and rerender the website
on change.

Client remains to be fully connected to the server, that is
still ahead in its API